+++ Changelog.md
@@ -1,5 +1,8 @@
 WhateverGreen Changelog
 =======================
+#### v1.3.8
+- Added `igfxfw=2` boot argument and `igfxfw` IGPU property to load Apple GuC firmware
+
 #### v1.3.7
 - Improved the maximum link rate fix: Now correct the value read from extended DPCD as well. (by @0xFireWolf)
 - Improved firmware loading handling on 10.15.4 (may fix booting issues on KBL+)